+++ pidx/data_handle/PIDX_idx_data_structs.h
@@ -241,7 +241,7 @@ struct idx_file_struct
 
   int compression_start_level;
   int compression_factor;
-  int compression_bit_rate;
+  float compression_bit_rate;
   unsigned long long chunk_size[PIDX_MAX_DIMENSIONS];
 
   int file_zero_merge;
